requirements-torch201-cuda117.txt
absl-py==1.4.0
accelerate @ git+https://github.com/huggingface/accelerate@c3d82d24e2f1dfb5c109e35354dd78744fe8b50c
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blobfile==2.0.2
cachetools==5.3.1
carvekit-colab==4.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.26.4
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.1.0
cubvh @ git+https://github.com/ashawkey/cubvh@d3e923c90e56974d39f57c2e2b9bf372d2e45870
cycler==0.11.0
dearpygui==1.9.1
debugpy==1.6.7
debugpy-run==1.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.17.1
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.2
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
freqencoder @ file:///nfs/code/gqian/img2mesh/freqencoder
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gradio==3.34.0
gradio_client==0.2.6
gridencoder @ file:///nfs/code/gqian/img2mesh/gridencoder
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
humanize==4.6.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipykernel==6.23.2
ipython==8.14.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.2
lightning-utilities==0.8.0
linkify-it-py==2.0.2
lit==16.0.5.post0
llvmlite==0.40.1rc1
loguru==0.7.0
lpips==0.1.4
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast/@335cfa6b33d785730a04283994214bed57884e87
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
pooch==1.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.8
PyMCubes==0.1.4
pymeshlab==2022.2.post4
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
raymarching @ file:///nfs/code/gqian/img2mesh/raymarching
regex==2023.6.3
rembg==2.0.43
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.25.1
setproctitle==1.3.2
shap-e @ git+https://github.com/openai/shap-e@db3f3b4c2f572b53fbe408d202e2445a1f656353
shencoder @ file:///nfs/code/gqian/img2mesh/shencoder
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
taming-transformers-rom1504==0.0.6
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-ema==0.3
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trimesh==3.22.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.2.0
urllib3==1.26.16
uvicorn==0.22.0
wandb==0.15.4
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.3.6
xatlas==0.0.7
yarl==1.9.2
zipp==3.15.0
